extract method to base controller
